+++ src/components/plot/svg-color-bar.tsx
@@ -110,6 +110,7 @@ export function SvgVColorBar({
   }
 
   ax = setAxisDirection(ax, 'y')
+
   const length = axisLength(ax)
 
   const colorStep = 1 / (steps - 1)